refactor to a service
