+++ app/services/component-lookup.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-cli-is-component/services/component-lookup';
