+++ app/services/is-component.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-cli-is-component/services/is-component';
